only one WindowProc for AssociatedObject
